add Visium Spatial transcriptomics voila dashboard
This example show how we can interactively explore high-dimensional and spatial transcriptomics data produced by the Visium assay from 10X Genomics. We are using Bqplot to display the spatial data and the interactive heatmap Clustergrammer2 to display the high-dimensional gene expression data. The views are linked and enable users to overlay gene expression or unbiased clustering results on the spatial view. We also allow transition to UMAP projection of data.