TrendPy  Time Series Regression with Python
TrendPy is a Python package tailored to easily perform powerful regressions with linear, polynomial and sinusoidal ansatzes. Further a three-parameter "free" regression with user ansatz is possible.

The TrendPy packages is object oriented and features classical aspects as returning the coefficients, coefficient of determination of the training data.

Further, a web application is provided, which allows to apply different regression models on data uploaded by the user, it extends some features of Excel's trendline.

This project is based on NumPy, SciPy Jupyter Notebook, Plotly, Voila and SymPy.